Refactor InspectionItemForm
